Convert render-item
