+++ app/utils/escape-reg-exp.ts
@@ -1,5 +1,5 @@
-/* eslint no-useless-escape: 0 */
-export default function (str) {
+/* eslint-disable no-useless-escape */
+export default function (str: string) {
   if (typeof str === 'string') {
     return str.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, '\\$&');
   }
